+++ DataMartBackend/src/main/java/ru/datamart/project/services/MesService.java
@@ -3,6 +3,8 @@ package ru.datamart.project.services;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import ru.datamart.project.customExceptions.CustomEntityNotFoundException;
+import ru.datamart.project.customExceptions.CustomInvalidRequestException;
 import ru.datamart.project.dto.BatchMesDto;
 import ru.datamart.project.dto.MesDto;
 import ru.datamart.project.models.DimBatchEntity;
@@ -22,9 +24,12 @@ public class MesService {
     private final DimBatchRepository batchRepo;
 
     public BatchMesDto getMesByBatchId(String batchId) {
+        if (batchId == null) {
+            throw new CustomInvalidRequestException("Укажите ID партии.");
+        }
         Optional<BatchMesDto> optional = mesRepo.getMesByBatchId(batchId);
         if (optional.isEmpty()) {
-            return null;
+            throw new CustomEntityNotFoundException("Данные MES для партии не были найдены.");
         }
         BatchMesDto dto = optional.get();
         if (dto.getStatus() != null) {
@@ -37,6 +42,9 @@ public class MesService {
     }
 
     public Optional<MesEntity> save(MesDto dto) {
+        if (dto.getRecordId() == null || dto.getBatchId() == null) {
+            throw new CustomInvalidRequestException("Укажите ID записи и партии.");
+        }
         DimBatchEntity batch = batchService.getOrCreate(dto);
         batch.setProcessStatus(dto.getProcessStatus());
         batch.setEndTime(dto.getEndTime());
@@ -60,10 +68,16 @@ public class MesService {
     }
 
     public Optional<MesEntity> get(String id) {
+        if (id == null) {
+            throw new CustomInvalidRequestException("Укажите ID записи.");
+        }
         return mesRepo.findById(id);
     }
 
     public void delete(String id) {
+        if (id == null) {
+            throw new CustomInvalidRequestException("Укажите ID записи.");
+        }
         mesRepo.deleteById(id);
         log.info("УДАЛЕНА ЗАПИСЬ MES");
     }
