+++ DataMartBackend/src/main/java/ru/datamart/project/services/NotificationService.java
@@ -3,6 +3,9 @@ package ru.datamart.project.services;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import ru.datamart.project.customExceptions.CustomEntityNotFoundException;
+import ru.datamart.project.customExceptions.InvalidCredentialsException;
+import ru.datamart.project.customExceptions.CustomInvalidRequestException;
 import ru.datamart.project.dto.SimpleWsMessageDto;
 import ru.datamart.project.models.NotificationEntity;
 import ru.datamart.project.models.NotificationSeverityEnum;
@@ -18,7 +21,14 @@ public class NotificationService {
     private final WebSocketService webSocketService;
     private final NotificationRepository notificationRepository;
 
+    public List<NotificationEntity> getActive() {
+        return notificationRepository.getActive();
+    }
+
     public long countAlarmsByBatchId(String batchId) {
+        if (batchId == null) {
+            throw new CustomInvalidRequestException("Укажите ID партии.");
+        }
         return notificationRepository.countAlarmsByBatchId(batchId);
     }
 
@@ -31,23 +41,31 @@ public class NotificationService {
         n.setEquipmentId(equipmentId);
         n.setSignalSource(signalSource);
         n.setSeverity(severity);
-        NotificationEntity saved = notificationRepository.save(n);
+        NotificationEntity saved;
+        try {
+            saved = notificationRepository.save(n);
+        } catch (Exception e) {
+            throw new InvalidCredentialsException("Не удалось создать уведомление.");
+        }
         webSocketService.sendNotificationsUpdate(new SimpleWsMessageDto("NOTIFICATION_create"));
         log.info("СОЗДАНО УВЕДОМЛЕНИЕ:\n" + saved);
         return saved;
     }
 
-    public List<NotificationEntity> getActive() {
-        return notificationRepository.getLastNotifications();
-    }
-
     public void markAsViewed(Long id) {
-        NotificationEntity n = notificationRepository.findById(id).orElseThrow();
+        if (id == null) {
+            throw new CustomInvalidRequestException("Укажите ID уведомления.");
+        }
+        NotificationEntity n = notificationRepository.findById(id)
+                .orElseThrow(() -> new CustomEntityNotFoundException("Уведомление с таким ID не было найдено."));
         n.setViewed(true);
         notificationRepository.save(n);
     }
 
     public void delete(Long id) {
+        if (id == null) {
+            throw new CustomInvalidRequestException("Укажите ID уведомления.");
+        }
         notificationRepository.deleteById(id);
     }
 }
